patch cache: recheck if somebody already added to cache
We hopefully serve most requests from the cache (when we acquired the
read lock). If not, multiple workers start to generate the patch. These
then acquire the write lock and add the patch to the cache. So war we
did that multiple times. While it did not race, it is not necessary.

Recheck and return the generated patch.

While unlikely, it could happen that multiple go routines calculate the
same patch until it hits the cache for the first time. Unlikely and not
worth to introduce a "already generating + poll" state or anything. This
is a very short period where this even could happen, then we serve from
the cache.