fix writing response
This is a nasty bug where the string got written with Fprintf. The
string (i.e., the patch that got generated) contains '%s' and other
format specifiers. So the patch that got written to back contained
"MISSING" arguments and therefore was invalid.

The string is the final patch, no need for formatted output.