fix compile test
This gcc call is just some test if the sent data is even valid C-header
code and not some garbage[1]. Recent gcc wants to generate a '.h.gch'
(precompiled headers) and that breaks if the output is '/dev/null'. Fix
by calling a syntax check only. Modern gcc has it, and the one in used
in the Dockerfile (Jammy as of writing) as well. Probably it exists for
decades anyways.

[1] okay, most C code also falls under that category...